[BACKLOG-21192] Removing pentaho.lang.DomNode and simplifying Scene
