+++ impl/client/src/test/javascript/pentaho/visual/scene/Base.spec.js
@@ -331,252 +331,6 @@ define([
       });
     });
 
-    describe("#removeChild(this) - child becoming a root", function() {
-
-      // impact in `root`, `view` and `vars`
-
-      var CustomView;
-      var view;
-      var parentScene;
-      var childScene;
-
-      beforeAll(function() {
-        CustomView = View.extend();
-      });
-
-      beforeEach(function() {
-        view = new CustomView();
-        parentScene = new Scene(null, view);
-        childScene = new Scene(parentScene);
-      });
-
-      it("should have #root be itself", function() {
-
-        parentScene.removeChild(childScene);
-
-        // ---
-
-        expect(childScene.root).toBe(childScene);
-        expect(childScene.parent).toBe(null);
-        expect(childScene.index).toBe(-1);
-      });
-
-      it("should have #view be the old view", function() {
-
-        parentScene.removeChild(childScene);
-
-        // ---
-
-        expect(childScene.view).toBe(view);
-      });
-
-      it("should have #vars have a null prototype", function() {
-
-        parentScene.removeChild(childScene);
-
-        // ---
-
-        expect(Object.getPrototypeOf(childScene.vars)).toBe(null);
-      });
-
-      it("should have #vars preserve its local variables", function() {
-
-        childScene.setVar("category", "Portugal");
-        childScene.setVar("measure", 6000);
-
-        var categoryVar = childScene.vars.category;
-        var measureVar = childScene.vars.measure;
-
-        // ---
-
-        parentScene.removeChild(childScene);
-
-        // ---
-
-        expect(childScene.vars.category).toBe(categoryVar);
-        expect(childScene.vars.measure).toBe(measureVar);
-      });
-
-      it("should have #vars of grand-children inherit from its #vars", function() {
-
-        var grandChild = new Scene(childScene);
-
-        // ---
-
-        parentScene.removeChild(childScene);
-
-        // ---
-
-        expect(Object.getPrototypeOf(grandChild.vars)).toBe(childScene.vars);
-      });
-
-      it("should have #vars of grand-children preserve their local variables", function() {
-
-        var grandChild = new Scene(childScene);
-
-        grandChild.setVar("series", "Lisbon");
-
-        var seriesVar = grandChild.vars.series;
-
-        // ---
-
-        parentScene.removeChild(childScene);
-
-        // ---
-
-        expect(grandChild.vars.series).toBe(seriesVar);
-      });
-    });
-
-    describe("#appendChild(this-a-root) - root becoming a child", function() {
-
-      // impact in `root`, `view` and `vars`
-
-      var CustomView;
-
-      var view1;
-      var view2;
-
-      var parentScene;
-      var childScene;
-      var grandChildScene;
-
-      beforeAll(function() {
-        CustomView = View.extend();
-      });
-
-      beforeEach(function() {
-        view1 = new CustomView();
-        view2 = new CustomView();
-
-        parentScene = new Scene(null, view1);
-        childScene = new Scene(parentScene);
-
-        // Initially a root.
-        grandChildScene = new Scene(null, view2);
-      });
-
-      it("should have #root be the new parent's root", function() {
-
-        childScene.appendChild(grandChildScene);
-
-        // ---
-
-        expect(grandChildScene.root).toBe(parentScene);
-      });
-
-      it("should have #parent be the new parent", function() {
-
-        childScene.appendChild(grandChildScene);
-
-        // ---
-
-        expect(grandChildScene.parent).toBe(childScene);
-        expect(grandChildScene.index).toBe(0);
-      });
-
-      it("should have #view be the new parent's view", function() {
-
-        childScene.appendChild(grandChildScene);
-
-        // ---
-
-        expect(grandChildScene.view).toBe(view1);
-      });
-
-      it("should have #vars have the parent's #vars as prototype", function() {
-
-        childScene.appendChild(grandChildScene);
-
-        // ---
-
-        expect(Object.getPrototypeOf(grandChildScene.vars)).toBe(childScene.vars);
-      });
-
-      it("should have #vars preserve its local variables", function() {
-
-        childScene.setVar("category", "Lisbon");
-
-        var categoryVar = childScene.vars.category;
-
-        childScene.appendChild(grandChildScene);
-
-        // ---
-
-        expect(grandChildScene.vars.category).toBe(categoryVar);
-      });
-    });
-
-    describe("#appendChild(this-a-child-of-another-parent) - changing parents", function() {
-
-      // impact in `root`, `view` and `vars`
-
-      var CustomView;
-
-      var view1;
-      var view2;
-
-      var parentScene1;
-      var parentScene2;
-      var childScene;
-
-      beforeAll(function() {
-        CustomView = View.extend();
-      });
-
-      beforeEach(function() {
-        view1 = new CustomView();
-        view2 = new CustomView();
-
-        parentScene1 = new Scene(null, view1);
-        parentScene2 = new Scene(null, view2);
-
-        childScene = new Scene(parentScene1);
-      });
-
-      it("should have #root be the new parent's root", function() {
-
-        parentScene2.appendChild(childScene);
-
-        // ---
-
-        expect(childScene.root).toBe(parentScene2.root);
-      });
-
-      it("should have #view be the new parent's view", function() {
-
-        parentScene2.appendChild(childScene);
-
-        // ---
-
-        expect(childScene.view).toBe(view2);
-      });
-
-      it("should have #vars have the new parent's #vars as prototype", function() {
-
-        parentScene2.appendChild(childScene);
-
-        // ---
-
-        expect(Object.getPrototypeOf(childScene.vars)).toBe(parentScene2.vars);
-      });
-
-      it("should have #vars preserve its local variables", function() {
-
-        childScene.setVar("category", "Lisbon");
-
-        var categoryVar = childScene.vars.category;
-
-        // ----
-
-        parentScene2.appendChild(childScene);
-
-        // ---
-
-        expect(childScene.vars.category).toBe(categoryVar);
-      });
-    });
-
     describe("#createFilter()", function() {
 
       var CustomView;
